+++ src/rules/property-blacklist/__tests__/index.js
@@ -84,9 +84,7 @@ testRule(rule, {
 testRule(rule, {
   ruleName,
 
-  config: [[
-    "/margin/",
-  ]],
+  config: ["/margin/"],
 
   accept: [ {
     code: "a { $margin: 0; }",